+++ changelog.md
@@ -10,6 +10,7 @@ For example:
 -->
 
 ### New Features
+
 * Add configs and docs for VSCode setup. [#854](https://gitlab.com/bramw/baserow/-/issues/854)
 * Added `in this week` filter [#569](https://gitlab.com/bramw/baserow/-/issues/954).
 * Allow users to use row id in the form redirect URL. [#871](https://gitlab.com/bramw/baserow/-/merge_requests/871)
@@ -24,6 +25,7 @@ For example:
 * Made it possible to import data into an existing table. [#342](https://gitlab.com/bramw/baserow/-/issues/342)
 * Tables can now be duplicated. [#961](https://gitlab.com/bramw/baserow/-/issues/961)
 * Introduced environment variable to disable Google docs file preview. [#1074](https://gitlab.com/bramw/baserow/-/issues/1074)
+* Made it possible to select the entire row via the row context menu. [#1061](https://gitlab.com/bramw/baserow/-/issues/1061)
 
 ### Bug Fixes